1期文件/day24 供应链系统/day24 供应链系统/代码/后端/dahe-3/requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.14.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
